Add spec for InvalidFieldType
