+++ lib/mongoid/errors.rb
@@ -15,6 +15,7 @@ require "mongoid/errors/invalid_config_option"
 require "mongoid/errors/invalid_dependent_strategy"
 require "mongoid/errors/invalid_field"
 require "mongoid/errors/invalid_field_option"
+require "mongoid/errors/invalid_field_type"
 require "mongoid/errors/invalid_find"
 require "mongoid/errors/invalid_includes"
 require "mongoid/errors/invalid_index"
